# contract/openapi.public.snapshot.json
{
  "components": {
    "schemas": {
...
      "BooksPageResponse": {
        "additionalProperties": false,
        "description": "Paginated bookshelf response.",
        "properties": {
          "global_mark_count": {
            "description": "Total number of marks across all books.",
            "title": "Global Mark Count",
            "type": "integer"
          },
          "items": {
            "description": "Bookshelf cards for the current page.",
            "items": {
              "$ref": "#/components/schemas/BookShelfCard"
            },
            "title": "Items",
            "type": "array"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination metadata for the bookshelf query."
          }
        },
        "required": [
          "items",
          "page_info",
          "global_mark_count"
        ],
        "title": "BooksPageResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Stable error envelope returned by every REST failure.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context for debugging or client-side branching.",
            "title": "Details"
          },
          "error_id": {
            "description": "Server-generated identifier for this specific error instance.",
            "title": "Error Id",
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation of the failure.",
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "description": "Whether retrying the same request may succeed.",
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "description": "HTTP status code returned with this error.",
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "error_id",
          "code",
          "message",
          "status",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "description": "Cursor pagination metadata.",
        "properties": {
          "has_more": {
            "description": "Whether another page of results is available.",
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "description": "The page size used for this response.",
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page, or null when there are no more results.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "limit",
          "has_more"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Deep Read Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/books": {
      "get": {
        "description": "Return the paginated bookshelf view.",
        "operationId": "books_index_api_books_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Books Index"
      }
    },
    "/api/books/{book_id}": {
      "get": {
        "description": "Return the result-view overview for one book.",
        "operationId": "book_detail_api_books__book_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "title": "Book Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Book Detail"
      }
    },
...
    "/api/books/{book_id}/analysis-log": {
      "get": {
        "description": "Return the latest technical log tail for one book analysis.",
        "operationId": "book_analysis_log_api_books__book_id__analysis_log_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "line_limit",
            "required": false,
            "schema": {
              "default": 120,
              "maximum": 400,
              "minimum": 20,
              "title": "Line Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Book Analysis Log"
      }
    },
...
  }
}